docs: update the post and Post to author and Author to be updated with the illustrated instance and logic
In the documentation of v7 alpha associations BelongsToMany association, as the section Association Adder (addX) starts, we are using two classes(models) Author and Book, but when creating an instance of the class(data item), Post.findByPk is being used instead of Author.findByPk, similarly post is being used instead of author to represent instance/s (data item/s). The issue is persisting in the subsequent section. This commits  aims towards tackling these issues. 

Thanks and warm regards, to the sequelize team and all the contributors for this great opportunity and product!